Split ModActions into separate controllers and types
The ModActions class file was just getting way too big and I learned a while back that a declared, non-exported type is global so it can be kept anywhere without consequence. Been meaning to clean up with that in mind since then.

Need to go through and adjust method paths in all commands to accomodate this change but I wanted to get it committed now in case I switch computers