Remove command class file class names
The default export takes care of it fine and they were just polluting my potential class name pool if I didn't want duplicates. The class name isn't required by the CommandLoader as long as the export has a key. Though the TS compiler will give these all the class name "default_1" but it's not really a big deal. The class